Casts have been relegated to one place for Weapon. Taking in a Vec<T> has been replaced with a &[T]

+++ src/weapon/mod.rs
@@ -10,6 +10,6 @@ pub struct Weapon<'s> {
     texture_loader: &'s TextureLoader,
     shadows: RectangleShape<'s>,
     shadows_id: Vec<i32>,
-    current_weapon: i32,
+    current_weapon: usize,
     mouse_fire: bool,
 }
